Introduce a new class, "State" that takes over some functionality previously handled by Grid.  State now owns:
* the solution array q
* the aux array
* aux_global
* t

Now a Solution object in general contains a list of States (rather than
a list of Grids), and each State has a grid attribute that is a
reference to the Grid it lives on.  In principle you could have multiple
States on a Grid, though I don't have a current use case.  This change
also required lots of internal changes to plotters/frametools.py in
VisClaw, but they should be invisible to the user.

Several things seem more natural now, and I'm fairly convinced that this
is an improvement.  But let me know if you disagree or think some
changes are for the worse.  It would be good to have a code review very
soon.

All tests and apps have been updated to use the new code.  This mostly
involved deleting commands that are no longer necessary, changing "grid"
to "state" in several places, and changing how aux is handled at the
boundaries.

Ghost-cell values of aux are now handled in exactly the same way as
ghost-cell values of q.  The user does not need to set them directly.
There is some potential for inefficiency here, especially in the
parallel case, so I plan to add a flag to indicate if the aux
variables are time-independent, in which case they will be calculated
just once.

In order to avoid using mbc in State, the q and aux DAs are initially
set up with stencil_width=0.  Later, in Solver.setup(), new DAs are
created using the correct stencil_width for the solver chosen.
Hopefully this inefficiency is not significant, since it only happens
once.