new fileio.ascii.read_t function consolidates function previously defined in 3 places
pyclaw.fileio.binary and pyclaw.solution now import from pyclaw.fileio.ascii
rather than having 3 versions that previously were not exactly consistent.
Also added except clause so that if file_format is not in the fort.t file,
it is backward compatible with output from running with a previous version.