Update and add tests for Pytest (#3)
* Add test for  nonunif advection and remove nose from advection

* Add tests for 1d advection with variable coeffs

Testing for some custom RK time integrators, relies on
dictionaries.

* rephrasing comment

* Add tests for 2D advection

Testing transverse solver.
I think there is too much numerical dissipation with such a coarse grid
to compare initial conditions with solution at final time.
Using a dictionary with expected solutions instead.

* Add test advection_annulus

* Update test for advection 1D with nonuniform grid

Using dictionary with expected solution instead of comparison between
initial condition and final solution.

* Remove unnecessary abs

* Update docstring

* Rename classes for consistency

* Add test Burgers 1d

* Fix bug relative path

* Update tests Euler 2D

Switch to pytest and use a single test file.
I am not using the tools from pyclaw.utils. Should we test them with
the examples? If so, maybe we should keep them in one or two examples.
I am skipping tests for shock_forward_step, so the HLLE RS with walls is
not tested. This example gives me some random NaNs that I am not able to
handle robustly.

* Add tests Euler 3D

Add tests for the other 2 examples for 3D Euler.

* Do not abort execution if there is no MPI

* Add init file for relative imports to work

Is there a reason to not treat this directory as a package?

* Better to not treat this one as package for the moment

* Update test Euler with gravity 3d

I didn't see why the test (with nose) was skipped
if scipy couldn't be found. I removed that requirement here.

The execution of both examples is not aborted now if mpi4py
is not available, just a warning is displayed.

The same .txt file with
data for the regression test is used.

The local Makefile is used to
compile the source with subprocess.

* Redirect output Sedov Euler 3d

* Skipping PeanoClaw tests for the moment

* Handle relative path for Pytest

* Fix import

* Disable output when testing

* Update test psystem 2d

Updated to work with pytest and removed dependence on pyclaw.util.

* Remove nosetest Euler w gravity 3D

* Add pytest SWEs 1D, remove nosetests

* Redirect output and remove nose calls

* Fix test class name

* Add pytests SWEs 2D

* Remove nose call

* Add pytests stegoton 1d

* Passing some paremeters as arguments for setup

* Add pytests traffic 1D

* Remove nose call

* Add expected solutions Burgers 1D

---------

Co-authored-by: Carlos Munoz Moncayo <munozmca@kl-26147.kaust.edu.sa>
Co-authored-by: Carlos Munoz Moncayo <munozmca@kl-26147.local>